requirements.txt
python-can
matplotlib